% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_block_radius_miles_to_this_dt.R
\name{add_block_radius_miles_to_dt}
\alias{add_block_radius_miles_to_dt}
\title{Create a new column in the data.table passed to this function (by reference)}
\usage{
add_block_radius_miles_to_dt(dt, area_sqmi)
}
\arguments{
\item{dt}{data.table}

\item{area_sqmi}{vector of numbers that are area in square miles.
If area were in square meters, it could be converted like this:

area_sqmi <-
census2020download:::area_sqmi_from_area_sqmeters(area_sqmeters)}
}
\value{
NULL. The side effect is that a new column is created by reference in
the data.table referred to here as dt, but
the data.table passed to this function
actually gets changed by reference in that calling environment
without this function needing to return anything.
}
\description{
Create a new column in the data.table passed to this function (by reference)
}
\details{
Note the table magically gets updated just by calling the function
-- Do not assign the function results to the object!
That would just turn the table into NULL.
}
\examples{
  passed_dt <- data.table::data.table(a = 1:10)
  x <- add_block_radius_miles_to_dt(passed_dt, 9001:9010)
  is.null(x)
  passed_dt

}
\seealso{
\code{\link[=block_radius_miles_from_area_sqmi]{block_radius_miles_from_area_sqmi()}}
}
