% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_radius_miles_from_area_sq_mi.R
\name{area_sqmi_from_area_sqmeters}
\alias{area_sqmi_from_area_sqmeters}
\title{helper - converts units}
\usage{
area_sqmi_from_area_sqmeters(area_sqmeters)
}
\arguments{
\item{area_sqmeters}{vector of areas in square meters}
}
\value{
vector of areas in square miles
}
\description{
helper - converts units
}
\details{
area in the data.table blockwts after initial cleaning was
called area and was in square meters
}
