% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_bgid2fips.R
\docType{data}
\name{bgid2fips}
\alias{bgid2fips}
\title{Census FIPS code of each US Census 2020 block group}
\description{
data.table created at run time for use in the EJAM package
(not bundled with this package).
}
\details{
See EJAM package for more info.

Created by \code{\link[=census2020_get_data]{census2020_get_data()}} and \code{\link[=census2020_save_datasets]{census2020_save_datasets()}}.

This package can create this data table but does not store it.

Columns:
\itemize{
\item bgid The unique integer ID of each block group (for joins with \link{blockwts}),
used as a more efficient ID than a 12-character FIPS code.
\item bgfips (12-character string of block group FIPS code)
}
}
