% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_radius_miles_from_area_sq_mi.R
\name{block_radius_miles_from_area_sqmi}
\alias{block_radius_miles_from_area_sqmi}
\title{simple formula to calculate radius given area
Just the square root of (area/pi)}
\usage{
block_radius_miles_from_area_sqmi(area_sqmi)
}
\arguments{
\item{area_sqmi}{vector of numbers that are the area of each block,
in square miles.
If area is available only in square meters, it can be converted like this:

area_sqmi <-
census2020download:::area_sqmi_from_area_sqmeters(area_sqmeters)}
}
\value{
vector of numbers same shape as input
}
\description{
simple formula to calculate radius given area
Just the square root of (area/pi)
}
\seealso{
\code{\link[=add_block_radius_miles_to_dt]{add_block_radius_miles_to_dt()}} \code{\link[=add_block_radius_miles_to_dt]{add_block_radius_miles_to_dt()}}
}
