% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockid2fips.R
\docType{data}
\name{blockid2fips}
\alias{blockid2fips}
\title{block id for each block fips code for US Census blocks}
\description{
BUT SEE EJAMblockdata package
For populated Census blocks only, from Census block dataset.
}
