% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_blockpoints.R
\docType{data}
\name{blockpoints}
\alias{blockpoints}
\title{lat/lon location of internal point of each Census 2020 block, used in EJAM package}
\description{
data.table for use in EJAM
}
\details{
See EJAM package for more info.

Created by \code{\link[=census2020_get_data]{census2020_get_data()}} and \code{\link[=census2020_save_datasets]{census2020_save_datasets()}}

This package can create this data table but does not store it.

Rows: 8,174,955

Columns:
\itemize{
\item blockid (for joins among tables \link{blockwts}, \link{blockpoints}, \link{quaddata}, \link{blockid2fips})
unique integer 1 through number of rows (blocks),
used as a more efficient ID than a 15-character FIPS code.
\item lat  Latitude decimal degrees
\item lon  Longitude decimal degrees
}
}
