% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockquadtree.R
\docType{data}
\name{blockquadtree}
\alias{blockquadtree}
\title{quad tree format data on locations of US Census 2020 blocks}
\description{
BUT SEE EJAMblockdata package
Used to quickly find which of these block points
are nearby a specified facility or site with circular buffer.

\preformatted{

  Formal class 'QuadTree' [package "SearchTrees"] with 7 slots
  ..@ ref      :<externalptr> 
  ..@ numNodes : int 5173
  ..@ dataNodes: int 3294
  ..@ maxDepth : int 7
  ..@ maxBucket: int 48705
  ..@ totalData: int 5806512
  ..@ dataType : chr "point"
  ..$ data  :<externalptr> 
  ..$ points: int 5806512
  ..$ year  : num 2020
  #  code in this package created quaddata and blockquadtree: 
  
  quaddata <- blockdata[ , .(BLOCK_X, BLOCK_Z, BLOCK_Y, blockid)]
  blockquadtree <- SearchTrees::createTree(quaddata, treeType = "quad", dataType = "point")
  
  }
}
