% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockwts.R
\docType{data}
\name{blockwts}
\alias{blockwts}
\title{2020 Decennial Census block group weights}
\description{
BUT SEE EJAMblockdata package
A data.table with    blockwt, and
blockid or blockfips,
bgid or bgfips
used by doaggregate() to summarize the block group score of the
average person in the buffer, as the population weighted mean
of blockgroup scores of all the blocks in the buffer.
}
\details{
This is drawn from 2020 Census dataset via census2020_get_data()

\preformatted{
  blockid2fips data.table will have blockid and blockfips columns
  
    - blockfips is the 15 character Census Bureau FIPS code for each Census block
  
    - blockid will be a unique integer ID 1 through total number of blocks, 
    to be used as a more efficient indexing than a FIPS code is.
  
  bgid2fips data.table will have bgid and bgfips columns
  
    - bgfips  is the 12 character Census Bureau FIPS code 
    for the parent blockgroup (i.e., the one containing the given block).
    and bgfips is used to join to a blockgroup dataset to get indicator scores.
  
    - bgid will be a unique integer ID 1 through total number of unique blockgroups,
    to be used as a more efficient indexing than a FIPS code is.
  
  blockwts data.table will have blockid, bgid, blockwt, block_radius_miles as columns
  
    - blockwt is the block's population weight, 
  calculated as the block population as a fraction of the parent blockgroup pop.
  Based on the latest decennial Census table of population count for each block.

  This table of weights can be used for calculating the 
  weighted mean or sum of each blockgroup score for a buffer 
  where only some blocks of any given blockgroup are in the buffer.
  The sum of weights from some blocks tells you 
  what fraction of its whole parent blockgroup's population count
  is in those blocks (the ones found inside a buffer, for example).
  
    - block_radius_miles is the effective radius of a block, 
      where pi times that radius squared is the block area in square miles.
        This gets used to adjust distances of sites at very short distances to the centroid.
 }

See \url{https://www.epa.gov/ejscreen}
}
