% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census2020_download.R
\name{census2020_download}
\alias{census2020_download}
\title{Download Census 2020 data files from FTP}
\usage{
census2020_download(folder = getwd(), mystates)
}
\arguments{
\item{folder}{Default is getwd()}

\item{mystates}{Character vector of 2 letter abbreviations, now optional - Default is 50 states + DC + PR here}
}
\value{
Effect is to download and save locally a number of data files.
}
\description{
Warning: Code is not tested.
Attempts to download data files for specified states
from the US Census Bureau's FTP site for Decennial Census file data.
see \url{https://www2.census.gov/programs-surveys/decennial/2020/technical-documentation/complete-tech-docs/summary-file/2020Census_PL94_171Redistricting_StatesTechDoc_English.pdf}
}
\examples{
\dontrun{
 library(census2020download)
 census2020_download('./census2020zip', mystates = c('MD', 'DC'))
 census2020_unzip('./census2020zip','./census2020out')
 c2 <- census2020_read(folder = './census2020out', mystates = c('MD', 'DC'))
 save(c2,file = 'census2020blocks.rdata')
 dim(c2)
 str(c2)
 head(c2)
 sum(c2$POP100)
 plot(c2$INTPTLON[substr(c2$GEOCODE,1,2)=='24'], c2$INTPTLAT[substr(c2$GEOCODE,1,2)=='24'], pch='.')
 }
}
\seealso{
\link{census2020_read} \link{census2020_unzip}
}
