% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census2020_download_islandareas.R
\name{census2020_download_islandareas}
\alias{census2020_download_islandareas}
\title{Download Census 2020 data files from FTP - For Island Areas VI GU MP AS}
\usage{
census2020_download_islandareas(
  folder = NULL,
  mystates = c("VI", "GU", "MP", "AS"),
  allstates = c("VI", "GU", "MP", "AS"),
  baseurl = "https://www2.census.gov/programs-surveys/decennial/2020/data/island-areas/",
  urlmiddle = "demographic-and-housing-characteristics-file/",
  zipnames_suffix = "2020.dhc.zip"
)
}
\arguments{
\item{folder}{Default is a tempdir. Folder and subfolder for data are created if they do not exist.}

\item{mystates}{Character vector of 2 letter abbreviations, optional,
\itemize{
\item Default is VI GU MP AS
}}

\item{allstates}{Default is VI GU MP AS}

\item{baseurl}{default is the FTP folder with the data}
}
\value{
Effect is to download and save locally a number of data files.
}
\description{
Download Census 2020 data files from FTP - For Island Areas VI GU MP AS
}
\examples{
\dontrun{
 # 
 }
 
}
\seealso{
\code{\link[=census2020_get_data]{census2020_get_data()}} \code{\link[=census2020_download]{census2020_download()}}
}
