% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census2020_get_data_islandareas.R
\name{census2020_get_data_islandareas}
\alias{census2020_get_data_islandareas}
\title{Download and clean up VI,GU,MP,AS Island Areas block data Census 2020 (for EJAM)}
\usage{
census2020_get_data_islandareas(
  folder = NULL,
  folderout = NULL,
  mystates = c("VI", "GU", "MP", "AS"),
  do_download = TRUE,
  do_unzip = TRUE,
  do_read = TRUE,
  do_clean = TRUE,
  overwrite = TRUE,
  sumlev = 150,
  ...
)
}
\arguments{
\item{folder}{For downloaded files. Default is a tempdir. Folder is created if it does not exist.}

\item{folderout}{path for assembled results files, default is what folder was set to.}

\item{mystates}{Default is c('VI','GU','MP','AS')}

\item{do_download}{whether to do \code{\link[=census2020_download_islandareas]{census2020_download_islandareas()}}}

\item{do_unzip}{whether to do \code{\link[=census2020_unzip_islandareas]{census2020_unzip_islandareas()}}}

\item{do_read}{whether to do \code{\link[=census2020_read_islandareas]{census2020_read_islandareas()}}}

\item{do_clean}{whether to do \code{\link[=census2020_clean_islandareas]{census2020_clean_islandareas()}}}

\item{overwrite}{passed to \code{\link[=census2020_download_islandareas]{census2020_download_islandareas()}}}

\item{sumlev}{set to 150, meaning blockgroup not block, since no block data for island areas in these files!}

\item{...}{passed to \code{\link[=census2020_read_islandareas]{census2020_read_islandareas()}}}
}
\value{
invisibly returns a data.table of US Census blocks with columns like
blockid lat lon pop area (area in square meters), or intermediate info
depending on do_read, do_clean, etc.
}
\description{
Download and clean up VI,GU,MP,AS Island Areas block data Census 2020 (for EJAM)
}
\details{
Table 1 with pop seems unavailable from this source for island areas.
}
\examples{
 \dontrun{
 x = census2020_get_data()
 y = census2020_get_data_islandareas()
 }
}
