% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census2020_read.R
\name{census2020_read}
\alias{census2020_read}
\alias{census2020_read_islandareas}
\title{Compile Census 2020 block data for all US states once downloaded and unzipped
used by \code{\link[=census2020_get_data]{census2020_get_data()}}}
\usage{
census2020_read(
  mystates = NULL,
  folder = NULL,
  sumlev = 750,
  filenumbers = 1,
  best_header_cols = c("LOGRECNO", "GEOCODE", "AREALAND", "AREAWATR", "POP100", "HU100",
    "INTPTLAT", "INTPTLON"),
  best_data_cols = paste0("P00", (20001:20011))
)
}
\arguments{
\item{mystates}{can be vector of 2-letter abbreviations of states}

\item{folder}{path}

\item{sumlev}{default is 750, for blocks (but those are not available for Island Areas)}

\item{filenumbers}{a vector with any or all of 1,2,3 --
default is file 1.
File01 has Tables P1 and P2. Those are the ones with total population count.
File02 has Tables P3, P4, and H1.
File03 has Table P5.}

\item{best_header_cols}{default is a few key columns like POP100, GEOCODE (fips), etc.}

\item{best_data_cols}{default is key race ethnicity fields}
}
\value{
data.frame of 1 row per block, for example
}
\description{
Compile Census 2020 block data for all US states once downloaded and unzipped
used by \code{\link[=census2020_get_data]{census2020_get_data()}}
}
\details{
Not extensively tested.

Attempts to read files already downloaded and unzipped, data files for specified states
from the US Census Bureau's website for Decennial Census file data.

Technical documentation of the files available from Census Bureau
and for list of possible fields etc.:
\itemize{
\item \href{https://www2.census.gov/programs-surveys/decennial/2020/technical-documentation/complete-tech-docs/summary-file/}{Technical Documentation folder}
\item \href{https://www2.census.gov/programs-surveys/decennial/2020/technical-documentation/complete-tech-docs/summary-file/2020Census_PL94_171Redistricting_StatesTechDoc_English.pdf}{Technical Documentation PDF for 2020 Census PL94 171 Redistricting States files}
\item \href{https://www2.census.gov/programs-surveys/decennial/2020/technical-documentation/complete-tech-docs/demographic-and-housing-characteristics-file-and-demographic-profile/2020census-demographic-and-housing-characteristics-file-and-demographic-profile-techdoc.pdf}{Technical Documentation PDF for 2020 Census details}
\item \href{https://www2.census.gov/programs-surveys/decennial/2020/technical-documentation/complete-tech-docs/demographic-and-housing-characteristics-file-and-demographic-profile/2020-dhc-table-matrix.xlsx}{Table Matrix xlsx }
\item \href{https://www2.census.gov/programs-surveys/decennial/2020/technical-documentation/island-areas-tech-docs/dhc/2020-iac-dhc-technical-documentation.pdf}{Island Areas PDF Tech. Doc. }
\item \href{https://www2.census.gov/programs-surveys/decennial/2020/data/island-areas/american-samoa/demographic-and-housing-characteristics-file/2020-iac-dhc-readme.pdf}{Island Areas PDF Readme for American Samoa}
}

\preformatted{
    for example:
 #  AREALAND      Area (Land)
 #  AREAWATR      Area (Water)

 #  BASENAME      Area Base Name
 #  NAME          Area Name-Legal/Statistical Area Description (LSAD) Term-Part Indicator
 #  FUNCSTAT      Functional Status Code
 #  GCUNI         Geographic Change User Note Indicator

 #  POP100        Population Count (100%)
 #  HU100         Housing Unit Count (100%)
 #  INTPTLAT      Internal Point (Latitude)
 #  INTPTLON      Internal Point (Longitude)

 File 1 has Table P1 and
  Table P2. HISPANIC OR LATINO, AND NOT HISPANIC OR LATINO BY RACE
 Universe: Total population
 P0020001 P0020002 P0020003 P0020004 P0020005 P0020006 P0020007 P0020008 P0020009 P0020010 P0020011 P0020012 P0020013 P0020014
 Total: P0020001
  Hispanic or Latino P0020002
 Not Hispanic or Latino: P0020003
  Population of one race: P0020004
  White alone P0020005
 Black or African American alone P0020006
 American Indian and Alaska Native alone P0020007
 Asian alone P0020008
 Native Hawaiian and Other Pacific Islander alone P0020009
 Some Other Race alone P0020010
 Population of two or more races: Population of two races: P0020011

 }

See the package \href{https://github.com/GL-Li/totalcensus}{totalcensus}
for another attempt to facilitate downloads of files of data.

See the \href{https://walker-data.com/tidycensus/}{tidycensus package} for using the API to request tabular data.
}
\examples{
\dontrun{
 # library(census2020download)
 census2020_download('./census2020zip', mystates = c('MD', 'DC'))
 census2020_unzip('./census2020zip','./census2020out')
 c2 <- census2020_read(folder = './census2020out', mystates = c('MD', 'DC'))
 dim(c2)
 str(c2)
 head(c2)
 sum(c2$POP100)
 plot(
   c2$INTPTLON[substr(c2$GEOCODE,1,2) == '24'],
   c2$INTPTLAT[substr(c2$GEOCODE,1,2) == '24'], pch = '.')
 c2$LOGRECNO <- NULL
 colnames(c2) <- census2020download::census_col_names_map$Rname[
    match(colnames(blocks2020), census2020download::census_col_names_map$ftpname)
    ]
 }


}
\seealso{
\code{\link[=census2020_download]{census2020_download()}} \code{\link[=census2020_unzip]{census2020_unzip()}} \code{\link[=census2020_get_data]{census2020_get_data()}}
}
