% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census2020_save_datasets.R
\name{census2020_save_datasets}
\alias{census2020_save_datasets}
\title{Create separate data.tables, and optionally save them in the EJAM package
This is just done when Census FIPS or bounds or points change.}
\usage{
census2020_save_datasets(
  x,
  metadata = NULL,
  add_metadata = TRUE,
  save_as_data_for_package = FALSE,
  overwrite = TRUE,
  keep_pop = FALSE
)
}
\arguments{
\item{x}{a single data.table called blocks that is from \code{\link[=census2020_get_data]{census2020_get_data()}},
with colnames blockfips, pop, area, lat, lon}

\item{metadata}{default is Census 2020 related, tries to use \href{https://ejanalysis.com/ejamdocs}{EJAM package}}

\item{add_metadata}{logical, whether to add EJAM-related metadata about date and version}

\item{save_as_data_for_package}{logical, whether to do \code{\link[usethis:use_data]{usethis::use_data()}} here}

\item{overwrite}{default is TRUE, but only relevant if usethis = TRUE}

\item{keep_pop}{set to TRUE to keep the blockpop (population counts) column,
but it is not used by EJAM except here to create the weights before it is dropped by default.}
}
\value{
A named list of these (large) data.tables for the EJAM package. They
are created at run time and are not bundled with this package; see each help
topic for its columns:
\itemize{
\item \link{bgid2fips}
\item \link{blockid2fips}
\item \link{blockpoints}
\item \link{blockwts}
\item \link{quaddata}
}
}
\description{
Create separate data.tables, and optionally save them in the EJAM package
This is just done when Census FIPS or bounds or points change.
}
\seealso{
\code{\link[=census2020_save_datasets]{census2020_save_datasets()}} creates individual data.tables,
after \code{\link[=census2020_get_data]{census2020_get_data()}} has done these:
\itemize{
\item \code{\link[=census2020_download]{census2020_download()}}
\item \code{\link[=census2020_unzip]{census2020_unzip()}}
\item \code{\link[=census2020_read]{census2020_read()}}
\item \code{\link[=census2020_clean]{census2020_clean()}}
}
}
