% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census2020_unzip_islandareas.R
\name{census2020_unzip_islandareas}
\alias{census2020_unzip_islandareas}
\title{like census2020_unzip() but for Island Areas}
\usage{
census2020_unzip_islandareas(
  zpathslocal,
  folderout,
  filegeogrep = "geo2020\\\\.dhc$",
  filedatagrep = "[^o]2020\\\\.dhc$"
)
}
\arguments{
\item{zpathslocal}{zip file paths with filename, unlike in \code{\link[=census2020_unzip]{census2020_unzip()}}}

\item{folderout}{see \code{\link[=census2020_unzip]{census2020_unzip()}}}

\item{filegeogrep}{see \code{\link[=census2020_unzip]{census2020_unzip()}}}

\item{filedatagrep}{see \code{\link[=census2020_unzip]{census2020_unzip()}}}
}
\value{
see \code{\link[=census2020_unzip]{census2020_unzip()}}
}
\description{
like census2020_unzip() but for Island Areas
}
