% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_col_names_map.R
\docType{data}
\name{census_col_names_map}
\alias{census_col_names_map}
\title{table that maps from a few Census variable names to short friendly variable names}
\description{
table that maps from a few Census variable names to short friendly variable names
}
