% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_census_col_names_map_as.R
\docType{data}
\name{census_col_names_map_as}
\alias{census_col_names_map_as}
\title{Table that maps some Census variable names to short friendly variable names - AMERICAN SAMOA}
\description{
Table that maps some Census variable names to short friendly variable names - AMERICAN SAMOA
}
\details{
data.frame:	19 obs. of  3 variables:

$ ftpname : chr  "LOGRECNO" "GEOCODE" "AREALAND" "AREAWATR" etc

$ rname   : chr  "logrecno" "blockfips" "arealand" "areawater" etc

$ longname: chr  "logical record number" "15 digit Census FIPS"  etc

Caution: The same indicator such as "nhba" is assigned
a different P00* number depending on which geographic area the data came from.
For example, "nhba" is

P0020006 in the US DHC data,

P0050005 in the VI DHC data, and

P0050027 in the Guam DHC data
}
\seealso{
\link{census_col_names_map}
}
