% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_radius_miles_from_area_sq_mi.R
\name{radius_miles_from_area_sqmi}
\alias{radius_miles_from_area_sqmi}
\title{helper - just calculates radius from area
Just the square root of (area/pi)}
\usage{
radius_miles_from_area_sqmi(area_sqmi)
}
\arguments{
\item{area_sqmi}{vector of areas of blocks in square miles.

If area is available only in square meters, it can be converted like this:

area_sqmeters <- 100000

area_sqmi <-
census2020download:::area_sqmi_from_area_sqmeters(area_sqmeters)

radius <- radius_miles_from_area_sqmi(area_sqmi)}
}
\value{
vector of numbers same shape as input
}
\description{
helper - just calculates radius from area
Just the square root of (area/pi)
}
\seealso{
\code{\link[=add_block_radius_miles_to_dt]{add_block_radius_miles_to_dt()}} \code{\link[=add_block_radius_miles_to_dt]{add_block_radius_miles_to_dt()}}
}
