% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census2020_clean.R
\name{census2020_clean}
\alias{census2020_clean}
\title{start to clean up download block data}
\usage{
census2020_clean(x, cols_to_keep = c("blockfips", "lat", "lon", "pop", "area"))
}
\arguments{
\item{x}{data from census2020_read()}

\item{cols_to_keep}{optional, which (renamed) columns to retain and return}
}
\value{
data.table with these columns by default:  blockfips lat lon pop area
}
\description{
Rename columns based on census2020_download::census_col_names_map
Drops columns not needed.
Returns it in data.table format.
}
\details{
area is in square meters

This is part of how the output of census2020_read()
can be cleaned up and split into smaller data files,
to be used in EJAM.
}
