% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census2020_clean.R
\name{census2020_clean}
\alias{census2020_clean}
\alias{census2020_clean_islandareas}
\title{Start to clean up download block data
Renames columns, drops most, calculates total area.}
\usage{
census2020_clean(
  x,
  cols_to_keep = c("blockfips", "lat", "lon", "pop", "pop100", "area")
)
}
\arguments{
\item{x}{data from \code{\link[=census2020_read]{census2020_read()}}}

\item{cols_to_keep}{optional, which (renamed) columns to retain and return}
}
\value{
data.table with these columns by default:  blockfips lat lon pop area
}
\description{
Start to clean up download block data
Renames columns, drops most, calculates total area.
}
\details{
Renames and drops columns based on names in
\link{census_col_names_map} and the parameter cols_to_keep,
but see census_col_names_map for what could be retained.

Returns table in data.table format.

area is in square meters and is sum of land and water areas.
}
