% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census2020_download_islandareas.R
\name{census2020_download_islandareas}
\alias{census2020_download_islandareas}
\title{Download Census 2020 data files from Census Bureau - For Island Areas VI GU MP AS}
\usage{
census2020_download_islandareas(
  folder = NULL,
  mystates = c("VI", "GU", "MP", "AS"),
  allstates = c("VI", "GU", "MP", "AS"),
  baseurl = "https://www2.census.gov/programs-surveys/decennial/2020/data/island-areas/",
  urlmiddle = "demographic-and-housing-characteristics-file/",
  zipnames_suffix = "2020.dhc.zip",
  overwrite = TRUE
)
}
\arguments{
\item{folder}{Default is a tempdir. Folder and subfolder for data are created if they do not exist.}

\item{mystates}{Character vector of 2 letter abbreviations, optional,
\itemize{
\item Default is VI GU MP AS
}}

\item{allstates}{Default is VI GU MP AS}

\item{baseurl}{default is the URL of the Census Bureau site's folder with the data}

\item{urlmiddle}{Default is empty for States info, but
for Island Areas, urlmiddle = "demographic-and-housing-characteristics-file/"}

\item{zipnames_suffix}{last part of the filenames Census provides - default should work}

\item{overwrite}{set to FALSE to skip download if filename already in folder,
but note it does not check if any existing file is corrupt/size zero/obsolete/etc.}
}
\value{
Effect is to download and save locally a number of data files.
}
\description{
Download Census 2020 data files from Census Bureau - For Island Areas VI GU MP AS
}
\examples{
\dontrun{
  blocks_islandareas <- census2020_download_islandareas()
 }

}
\seealso{
\code{\link[=census2020_get_data]{census2020_get_data()}} \code{\link[=census2020_download]{census2020_download()}}
}
