% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census2020_get_data.R
\name{census2020_get_data}
\alias{census2020_get_data}
\title{Download and clean up US States,DC,PR, or Island Areas block data Census 2020 (for EJAM)}
\usage{
census2020_get_data(
  mystates = c(state.abb, "DC", "PR"),
  folder = NULL,
  folderout = NULL,
  do_download = TRUE,
  do_unzip = TRUE,
  do_read = TRUE,
  do_clean = TRUE,
  overwrite = TRUE,
  sumlev = 750
)
}
\arguments{
\item{mystates}{default is DC, PR, and the 50 states -- lacks
the island areas c('VI','GU','MP','AS') --
but \code{\link[=census2020_get_data]{census2020_get_data()}} can in some cases handle a mix of States/DC/PR
and/or island areas via helper function \code{\link[=census2020_get_data_islandareas]{census2020_get_data_islandareas()}},
returning either type of data, or a combined data.table if both are requested.
But block resolution is not available from these files for island areas,
so default for those is to get block groups,
which would not make sense to mix with blocks for states.}

\item{folder}{For downloaded files. Default is a tempdir. Folder is created if it does not exist.}

\item{folderout}{path for assembled results files, default is what folder was set to.}

\item{do_download}{whether to do \code{\link[=census2020_download]{census2020_download()}}, e.g., to just do subsequent steps if that one step was already done, but depends on temp folder, etc. so easier to just download again (default).}

\item{do_unzip}{whether to do \code{\link[=census2020_unzip]{census2020_unzip()}}}

\item{do_read}{whether to do \code{\link[=census2020_read]{census2020_read()}}}

\item{do_clean}{whether to do \code{\link[=census2020_clean]{census2020_clean()}}}

\item{overwrite}{passed to \code{\link[=census2020_download]{census2020_download()}}}

\item{sumlev}{Generally should not be changed from defaults.
Value of 750 means blocks, the only option likely to work here.
150 would mean blockgroups as for Island Areas since they lack block data here.
140 is tracts, 40 and 50 are State and County.
If mystates are Island Areas, this function uses 150 instead of 750.
But a mix of resolutions would not really make sense.}
}
\value{
invisibly returns a data.table of US Census blocks with columns like
blockid lat lon pop area (area in square meters), or just intermediate info
depending on do_read, do_clean, etc.
}
\description{
Download and clean up US States,DC,PR, or Island Areas block data Census 2020 (for EJAM)
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  To create certain data tables used by the EJAM package,
  which provides reports for EJSCREEN,
  EJAM relied on the census2020_download package, and
  used scripts like EJAM/data-raw/datacreate_ . . . .R
  to do something like this:

  blocks <- [census2020_get_data()] # default excludes Island Areas
  mylist <- [census2020_save_datasets(blocks)]

  bgid2fips    = mylist$bgid2fips
  blockid2fips = mylist$blockid2fips
  blockpoints  = mylist$blockpoints
  blockwts     = mylist$blockwts
  quaddata     = mylist$quaddata
}\if{html}{\out{</div>}}

For technical details on the files downloaded and tables and variables,
see the detailed references in the help for \code{\link[=census2020_read]{census2020_read()}}.
}
\examples{
 \dontrun{
 x = census2020_get_data(mystates="RI")

 y = [census2020_get_data()] # All States/DC/PR at block resolution
 z = [census2020_get_data_islandareas()] # VI,GU,MP,AS at blockgroup scale
 }
}
\seealso{
\code{\link[=census2020_save_datasets]{census2020_save_datasets()}} creates individual data.tables,
after \code{\link[=census2020_get_data]{census2020_get_data()}} has done these:
\itemize{
\item \code{\link[=census2020_download]{census2020_download()}}
\item \code{\link[=census2020_unzip]{census2020_unzip()}}
\item \code{\link[=census2020_read]{census2020_read()}}
\item \code{\link[=census2020_clean]{census2020_clean()}}
. and see \code{\link[=census2020_get_data_islandareas]{census2020_get_data_islandareas()}}
}
}
