% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census2020_get_data.R
\name{census2020_get_data}
\alias{census2020_get_data}
\title{Download and clean up US States,DC,PR block data Census 2020 (for EJAM)}
\usage{
census2020_get_data(
  folder = NULL,
  folderout = folder,
  mystates = c(state.abb, "DC", "PR"),
  do_download = TRUE,
  do_unzip = TRUE,
  do_read = TRUE,
  do_clean = TRUE,
  ...
)
}
\arguments{
\item{folder}{For downloaded files. Default is a tempdir. Folder is created if it does not exist.}

\item{folderout}{path for assembled results files, default is what folder was set to.}

\item{mystates}{default is DC, PR, and the 50 states -- lacks VI/GU/MP/AS}

\item{do_download}{whether to do census2020_download()}

\item{do_unzip}{whether to do census2020_unzip()}

\item{do_read}{whether to do census2020_read()}

\item{do_clean}{whether to do census2020_clean()}

\item{...}{passed to census2020_read()}
}
\value{
invisibly returns a data.table of US Census blocks with columns like
blockid lat lon pop area (area in square meters)
}
\description{
Download and clean up US States,DC,PR block data Census 2020 (for EJAM)
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  To create the individual data tables used by EJAM,
  
  blockwts, blockpoints, quaddata, blockid2fips, and bgid2fips,

  try something like
  
  
  blocks <- census2020_get_data()
  
  mylistoftables <- census2020_save_datasets(blocks)
  
  
  where mylistoftables will be
  
  
  list(bgid2fips = bgid2fips,
  
  blockid2fips = blockid2fips,
  
  blockpoints = blockpoints,
  
  blockwts = blockwts,
  
  quaddata = quaddata )
  
  
  See ?census2020_save_datasets
}\if{html}{\out{</div>}}
}
\seealso{
\code{\link[=census2020_get_islandareas]{census2020_get_islandareas()}}, \code{\link[=census2020_save_datasets]{census2020_save_datasets()}} which creates individual data.tables
census2020_get_data() just does \code{\link[=census2020_download]{census2020_download()}} then \code{\link[=census2020_unzip]{census2020_unzip()}}
then \code{\link[=census2020_read]{census2020_read()}} then \code{\link[=census2020_clean]{census2020_clean()}}
}
