% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census2020_get_data.R
\name{census2020_get_data}
\alias{census2020_get_data}
\title{Download and clean up block data census 2020}
\usage{
census2020_get_data(
  folder = "~/../Downloads/census2020zip",
  folderout = "~/../Downloads/census2020out",
  mystates = c(state.abb, "DC"),
  ...
)
}
\arguments{
\item{folder}{path for downloaded files}

\item{folderout}{path for assembled results files}

\item{mystates}{default is DC and the 50 states, not PR not VI/GU/MU/AS}

\item{...}{passed to census2020_read()}
}
\value{
a data.table of US Census blocks with columns like
blockid lat lon pop area (area in square meters I think)
}
\description{
Download and clean up block data census 2020
}
