% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census2020_get_data.R
\name{census2020_get_data}
\alias{census2020_get_data}
\title{Download and clean up block data census 2020}
\usage{
census2020_get_data(
  folder = "~/../Downloads/census2020zip",
  folderout = "~/../Downloads/census2020out",
  mystates = c(state.abb, "PR", "DC"),
  do_download = TRUE,
  do_unzip = TRUE,
  do_read = TRUE,
  do_clean = TRUE,
  ...
)
}
\arguments{
\item{folder}{path for downloaded files, or you could try setting it to tempdir()}

\item{folderout}{path for assembled results files, or you could try setting it to tempdir()}

\item{mystates}{default is DC and the 50 states -- lacks PR and VI/GU/MU/AS
and probably would need those from a different source}

\item{do_download}{whether to do census2020_download()}

\item{do_unzip}{whether to do census2020_unzip()}

\item{do_read}{whether to do census2020_read()}

\item{do_clean}{whether to do census2020_clean()}

\item{...}{passed to census2020_read()}
}
\value{
invisibly returns a data.table of US Census blocks with columns like
blockid lat lon pop area (area in square meters)
}
\description{
Download and clean up block data census 2020
}
