% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census2020_get_data_islandareas.R
\name{census2020_get_data_islandareas}
\alias{census2020_get_data_islandareas}
\title{Download and clean up VI,GU,MP,AS Island Areas block data Census 2020 (for EJAM)}
\usage{
census2020_get_data_islandareas(
  folder = NULL,
  mystates = c("VI", "GU", "MP", "AS"),
  do_download = TRUE,
  do_unzip = TRUE,
  do_read = TRUE,
  do_clean = TRUE
)
}
\arguments{
\item{folder}{Default is a tempdir. Folder is created if it does not exist.}

\item{mystates}{Default is c('VI','GU','MP','AS')}

\item{do_download}{whether to do census2020_download()}

\item{do_unzip}{whether to do census2020_unzip()}

\item{do_read}{whether to do census2020_read()}

\item{do_clean}{whether to do census2020_clean()}
}
\value{
invisibly returns list of tables
}
\description{
Download and clean up VI,GU,MP,AS Island Areas block data Census 2020 (for EJAM)
}
