% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census2020_get_data_islandareas.R
\name{census2020_get_data_islandareas}
\alias{census2020_get_data_islandareas}
\title{Download and clean up VI,GU,MP,AS Island Areas block data Census 2020 (for EJAM)}
\usage{
census2020_get_data_islandareas(
  mystates = c("VI", "GU", "MP", "AS"),
  folder = NULL,
  folderout = NULL,
  do_download = TRUE,
  do_unzip = TRUE,
  do_read = TRUE,
  do_clean = TRUE,
  overwrite = TRUE,
  sumlev = 150
)
}
\arguments{
\item{mystates}{Default is c('VI','GU','MP','AS')
data were at https://www2.census.gov/programs-surveys/decennial/2020/data/island-areas/
Also see page a-22 in https://www2.census.gov/programs-surveys/decennial/2020/technical-documentation/complete-tech-docs/supplemental-demographic-and-housing-characteristics-file/2020census-supplemental-dhc-techdoc.pdf}

\item{folder}{For downloaded files. Default is a tempdir. Folder is created if it does not exist.}

\item{folderout}{path for assembled results files, default is what folder was set to.}

\item{do_download}{whether to do \code{\link[=census2020_download_islandareas]{census2020_download_islandareas()}}}

\item{do_unzip}{whether to do \code{\link[=census2020_unzip_islandareas]{census2020_unzip_islandareas()}}}

\item{do_read}{whether to do \code{\link[=census2020_read_islandareas]{census2020_read_islandareas()}}}

\item{do_clean}{whether to do \code{\link[=census2020_clean_islandareas]{census2020_clean_islandareas()}}}

\item{overwrite}{passed to \code{\link[=census2020_download_islandareas]{census2020_download_islandareas()}}}

\item{sumlev}{set to 150, meaning blockgroup not block.

However, note this from Census Bureau:
"With this release of the 2020 IAC Demographic and Housing Characteristics Summary File,
the Census Bureau provides additional demographic
and housing characteristics for the Island Areas
down to the block, block group, and census tract levels."
Despite this, it appears that H1 (housing) table data are provided
at block resolution, but P1 (population count) is only at block group, tract, etc.
according to page 3 of the \href{https://www2.census.gov/programs-surveys/decennial/2020/technical-documentation/island-areas-tech-docs/dhc/2020-iac-dhc-technical-documentation.pdf}{Island Areas Tech. Doc.}}
}
\value{
invisibly returns a data.table of US Census units with columns like
the id, lat lon pop area (area in square meters), or intermediate info
depending on do_read, do_clean, etc.
}
\description{
Download and clean up VI,GU,MP,AS Island Areas block data Census 2020 (for EJAM)
}
\details{
Table 1 with pop seems unavailable from this source for island areas
if trying to use the same approach to reading files as done for the US States.

For Island areas, see https://www2.census.gov/programs-surveys/decennial/2020/technical-documentation/island-areas-tech-docs/dhc/2020-iac-dhc-technical-documentation.pdf
or e.g., (https://www2.census.gov/programs-surveys/decennial/2020/data/island-areas/american-samoa/demographic-and-housing-characteristics-file/2020-iac-dhc-readme.pdf)

For technical details on the files downloaded and tables and variables,
see the detailed references in the help for \code{\link[=census2020_read]{census2020_read()}}.
}
\examples{
 \dontrun{
 x = [census2020_get_data()] # States/DC/PR at block resolution
 y = [census2020_get_data_islandareas()] # VI,GU,MP,AS (at blockgroup scale)
 }
}
\keyword{internal}
