% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census2020_save_datasets.R
\name{census2020_save_datasets}
\alias{census2020_save_datasets}
\title{Create each separate data.table, and optionally save for use in a package
This was just done once, to create datasets for the EJAM package}
\usage{
census2020_save_datasets(
  x,
  metadata = NULL,
  add_metadata = TRUE,
  save_as_data_for_package = FALSE,
  overwrite = TRUE
)
}
\arguments{
\item{x}{a single data.table called blocks that is from \code{\link[=census2020_get_data]{census2020_get_data()}},
with colnames blockfips, pop, area, lat, lon}

\item{metadata}{default is Census 2020 related, tries to use EJAM package}

\item{add_metadata}{logical, whether to add metadata about date and version}

\item{save_as_data_for_package}{logical, whether to do \code{\link[usethis:use_data]{usethis::use_data()}} here}

\item{overwrite}{default is TRUE, but only relevant if usethis = TRUE}
}
\value{
A named list of these huge data.tables for the EJAM package:
\itemize{
\item bgid2fips
\item blockid2fips
\item blockpoints
\item blockwts
\item quaddata
}
}
\description{
Create each separate data.table, and optionally save for use in a package
This was just done once, to create datasets for the EJAM package
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  To create the individual data tables used by EJAM, 
  
  blockwts, blockpoints, quaddata, blockid2fips, and bgid2fips,
  
  from within the EJAM source package root folder, try something like 
  
  
  blocks <- census2020_get_data()
  
  mylistoftables <- census2020_save_datasets(blocks, 
    save_as_data_for_package=TRUE, overwrite=TRUE)

  
  See ?census2020_get_datasets
  
  and see create_quaddata() in EJAM pkg. 
}\if{html}{\out{</div>}}
}
