% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census2020_unzip.R
\name{census2020_unzip}
\alias{census2020_unzip}
\title{unzip Census 2020 zipped files already downloaded}
\usage{
census2020_unzip(folder = ".", folderout = folder, filenumbers = 1, mystates)
}
\arguments{
\item{folder}{path to where zip files are, default is working directory}

\item{folderout}{path to where you want to put files, created if does not exist}

\item{filenumbers}{a vector with any or all of 1,2,3 --
default is file 1.
File01 has Tables P1 and P2, which have population counts by race ethnicity.
File02 has Tables P3, P4, and H1.
File03 has Table P5.}

\item{mystates}{optional vector of 2letter state abbreviations for which to unzip}
}
\value{
Vector of filenames of unzipped contents
}
\description{
Warning: Code is not tested.
Attempts to read files already downloaded and unzipped, data files for specified states
from the US Census Bureau's FTP site for Decennial Census file data.
see \url{https://www2.census.gov/programs-surveys/decennial/2020/technical-documentation/complete-tech-docs/summary-file/2020Census_PL94_171Redistricting_StatesTechDoc_English.pdf}
}
\examples{
\dontrun{
 library(census2020download)
 census2020_download('./census2020zip', mystates = c('MD', 'DC'))
 census2020_unzip('./census2020zip','./census2020out')
 c2 <- census2020_read(folder = './census2020out', mystates = c('MD', 'DC'))
 save(c2,file = 'census2020blocks.rdata')
 dim(c2)
 str(c2)
 head(c2)
 sum(c2$POP100)
 plot(c2$INTPTLON[substr(c2$GEOCODE,1,2)=='24'], c2$INTPTLAT[substr(c2$GEOCODE,1,2)=='24'], pch='.')
 }
}
\seealso{
\link{census2020_download} \link{census2020_read}
}
