% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census2020download-package.R
\docType{package}
\name{census2020download}
\alias{census2020download}
\alias{census2020download-package}
\title{2020 Census Blocks and Tools to Download from FTP}
\description{
Basic functions for downloading from FTP,
unzipping, reading the 2020 Census data for some or all US States
into a single data.table, and splitting into a few tables
for use in the EJAM package.
}
\details{
This package is just a set of helper functions used by the EJAM package.

It creates block datasets (and some blockgroup tables) for use in the EJAM package.
It has basic functions for downloading from FTP,
unzipping, reading the 2020 Census data for some or all US States
into data.table format.

It can retain a few key variables like
\itemize{
\item lat and lon of block internal point
\item FIPS codes
\item population count or weight
\item area (or effective radius)
}

For more information see
\itemize{
\item \href{https://ejanalysis.github.io/census2020download/reference/census2020download.html}{census2020download package documentation}
\item \href{https://github.com/ejanalysis/census2020download}{census2020download package github repository}
\item \href{https://usepa.github.io/EJAM}{EJAM package}{.uri}
\item \href{https://www.census.gov/programs-surveys/acs/geography-acs/geography-boundaries-by-year.html}{Census vs ACS geos info}
}

\emph{Key functions and data.tables created include}
\itemize{
\item \code{\link[=census2020_get_data]{census2020_get_data()}} Download/ Unzip/ Read/ Clean data on
all US Census blocks
\item \code{\link[=census2020_save_datasets]{census2020_save_datasets()}} Use the data to create separate data.tables,
and save for use in the EJAM package.
}

\emph{Key data.table objects created:}
\itemize{
\item \link{blockid2fips} - data.table with FIPS code to blockid lookup
\item \link{blockpoints}  - data.table with latitude and longitude of internal points
\item \link{quaddata} - data.table with xyz format locations of blocks,
used to create spatial index of blocks in the EJAM package.
\item \link{blockwts} - data.table with Census 2020 population-based weight
as fraction of parent block group population, and size of block
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ejanalysis.github.io/census2020download/}
  \item \url{https://ejanalysis.github.io/census2020download/index.html}
}

}
\author{
\strong{Maintainer}: Mark Corrales \email{corrales.mark@epa.gov}

}
