% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_census_col_names_map.R
\docType{data}
\name{census_col_names_map}
\alias{census_col_names_map}
\title{Table that maps some Census variable names to short friendly variable names}
\description{
Table that maps some Census variable names to short friendly variable names
}
\details{
str(census_col_names_map)

data.frame:	19 obs. of  3 variables:

$ ftpname : chr  "LOGRECNO" "GEOCODE" "AREALAND" "AREAWATR" ...

$ Rname   : chr  "logrecno" "blockfips" "arealand" "areawater" ...

$ longname: chr  "logical record number" "15 digit Census FIPS"  etc.
}
