% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quaddata.R
\docType{data}
\name{quaddata}
\alias{quaddata}
\title{xyz location of internal point of each Census 2020 block,
for spatial index in EJAM}
\description{
data.table for use in EJAM
}
\details{
See EJAM package for more info.

Created by \code{\link[=census2020_get_data]{census2020_get_data()}} and \code{\link[=census2020_save_datasets]{census2020_save_datasets()}}

This package can create this data table but does not store it.

Rows: 8,174,955

Columns:
\itemize{
\item blockid (for joins among tables \link{blockwts}, \link{blockpoints}, \link{quaddata}, \link{blockid2fips})
unique integer 1 through number of rows (blocks),
used as a more efficient ID than a 15-character FIPS code.
\item x, y, z coordinates as shown below:
}

\preformatted{

  str(quaddata)

  Classes ‘data.table’ and 'data.frame':	8174955 obs. of  4 variables:

   $ BLOCK_X: num  205 205 204 204 204 etc
   $ BLOCK_Z: num  2125 2125 2125 2126 2126 etc
   $ BLOCK_Y: num  -3334 -3334 -3334 -3334 -3334 etc
   $ blockid: int  1 2 3 4 5 6 7 8 9 10 etc

     - attr(*, ".internal.selfref")=<externalptr>
     - attr(*, "sorted")= chr "blockid"
     - attr(*, "census_version")= num 2020
     - attr(*, "acs_version")= chr "2016-2020"
     - attr(*, "ejscreen_version")= chr "2.1"
     - attr(*, "ejscreen_releasedate")= chr "September 2022"
}
}
